.Rbuildignore the files that R doesn't care about
While it's important to have the pkgdown configuration file, the Rstudio
project settings, and the source for the README.md in the repo, R
doesn't care about these files and will produce NOTEs about them. List
them in .Rbuildignore so that they don't become a part of the source
package.